==> AkbarEtAl_LythrumFloralDisplay/1_ConceptualModel.Rmd <==
---
title: 'Flowering Schedule: Conceptual Model'
date: "`r Sys.Date()`"
output:
  pdf_document: default
  html_document: default
---

```{r setup}

knitr::opts_chunk$set(echo = TRUE)

library(knitr) # Kable for formatted tables
library(dplyr) # Data organizing
library(sn) # Generate normal distributions with skew

# Plotting
library(ggplot2)
library(ggpubr) 
theme_custom<-function(base_size = 24, base_family = ""){
  theme_bw(base_size = base_size, base_family = base_family) %+replace% 
    theme(axis.text = element_text(size = rel(0.8),colour="black"),
          axis.ticks = element_line(colour = "black"), 
          legend.position = "none",
          panel.border = element_rect(fill = NA, colour = NA), 
          panel.grid.major = element_line(colour = NA,size = 0), 
          panel.grid.minor = element_line(colour = NA,size = 0), 
          axis.line = element_line(colour ="black")
    )
}
setwd("./")
```


# Individual Flowering Schedules

Data setup

```{r ind_data}
t<-c(-100:100)/10 # Time (x-axis)
Norm<-data.frame(t=t,P_t=dnorm(t,sd=2.5),Shape="Normal")
Lepto<-data.frame(t=t,P_t=dt(t,df=1),Shape="Leptokurtic")
Even<-data.frame(t=t,P_t=rep(1/length(t),length(t)),Shape="Even")
#Pos<-data.frame(t=t,P_t=dnorm(t-20,sd=100),Shape="Positive Skew")
#Neg<-data.frame(t=t,P_t=dnorm(c(-580:20),sd=100),Shape="Negative Skew")
#qplot(x=t,y=dsn(t,xi=-8,omega=5,alpha=5,tau=1/100))
Pos<-data.frame(t=t,P_t=dsn(t,xi=-9,omega=3,alpha=5,tau=1/100),
                Shape="Positive Skew")
Neg<-data.frame(t=t,P_t=dsn(t,xi=9,omega=3,alpha=-5,tau=1/100),
                Shape="Negative Skew")

PDat<-rbind(Norm,Lepto,Even,Pos,Neg)
PDat$P_t[PDat$t==min(t)]<-0
PDat$P_t[PDat$t==max(t)]<-0
```

Make shape a factor and rearrange order for plotting

```{r ind_shape}
PDat$Shape<-factor(PDat$Shape, levels=c(
  "Normal","Leptokurtic","Even","Positive Skew","Negative Skew"))
```

Rescale:

1. Time from 0 to 1
2. Each curve to produce same total number of flowers (i.e. area=1)

```{r ind_scale}
PDat<-PDat %>% 
  group_by(Shape) %>%
  mutate(t=t/max(t),
         P_t=P_t/sum(P_t))
```

Make a wide version for plotting curves independently 

```{r ind_wide}
WDat <- pivot_wider(PDat, names_from = Shape, values_from = P_t)
```


## Visualization

### One curve per figure 

```{r ind_single}
ggplot(aes(x=t,y=P_t, colour="blue"),data=WDat) +
  geom_line(aes(lty=Shape),linewidth=1) + 
  ylab(expression(P[t])) +
  theme(axis.title.y = element_text(angle = 0),
        axis.ticks = element_blank(),
        axis.text.x = element_blank(),
        axis.text.y = element_blank())
```


### All curves in one figure 

```{r ind_all}
Fig1A <- ggplot(aes(x=t,y=P_t,colour=Shape),data=PDat) +
  geom_line(aes(lty=Shape),linewidth=1) + 
  ylab(expression(P[t])) +
  theme_custom() +
  theme(axis.title.y = element_text(angle = 90),
        axis.ticks = element_blank(),
        axis.text.x = element_blank(),
        axis.text.y = element_blank()) 
Fig1A
```

```{r ind_export}
pdf("./Figures/Box1/Fig1A.pdf",width=10,height=6)
Fig1A
dev.off()
```

**Figure 1A?**

## Flowering Schedule Summary Statistics

NOTE: These calculations took me way too long to figure out, after making many errors

```{r ind_sum}
SumStats<-PDat %>% 
  group_by(Shape) %>%
  summarize(Total_F = sum(P_t),
            Alpha=min(t), Omega=max(t), Duration=max(t)/max(t),
            Mu=round(
              sum(t*P_t),
              1),
            Var=round(
              sum((t-Mu)^2*P_t),
              3),
            CS=round(
              sum((t-Mu)^3*P_t)/sqrt(Var)^3,
              1),
            CK=round(
              sum((t-Mu)^4*P_t)/sqrt(Var)^4-3,
              1)
            )
```

Summary statistics for each curve:

```{r ind_table}
kable(t(SumStats), format = "markdown")
```

# Aggregate Population

Calculate aggregate flowers over time

```{r ag_data}
AgDat<-data.frame(t=t,
                  F_t=Norm$P_t + Even$P_t +
                    Lepto$P_t +
                    Pos$P_t + Neg$P_t)
AgDat$P_t<-AgDat$F_t/sum(AgDat$F_t)
AgDat$P_t[AgDat$t==min(t)]<-0
AgDat$P_t[AgDat$t==max(t)]<-0
```

Rescale time and P_t

```{r ag_scale}
AgDat<-AgDat %>% 
  mutate(t=t/max(t),
         P_t=P_t/sum(P_t))
```

## Visualization

```{r ag_plot}
Fig1B <- ggplot(aes(x=t,y=P_t),data=AgDat) +
  geom_line(linewidth=1) + 
  ylab(expression(P[t])) +
  theme_custom() +
  theme(axis.title.y = element_text(angle = 90),
        axis.ticks = element_blank(),
        axis.text.x = element_blank(),
        axis.text.y = element_blank())
Fig1B
```

```{r ag_export}
pdf("./Figures/Box1/Fig1B.pdf",width=10,height=6)
Fig1B
dev.off()
```

**Figure 1B?**

Or combine in a single figure?

```{r onefig}
FullDat<-AgDat %>%
  select(t,P_t) %>%
  mutate(Shape="Aggregate")
FullDat<-rbind(PDat,FullDat)
ggplot(aes(x=t,y=P_t,colour=Shape),data=FullDat) +
  geom_line(aes(lty=Shape),linewidth=1) + 
  ylab(expression(P[t])) +
  theme(axis.title.y = element_text(angle = 0),
        axis.ticks = element_blank(),
        axis.text.x = element_blank(),
        axis.text.y = element_blank())
```

**Figure 1.** Description...

## Flowering Schedule Summary Statistics

```{r ag_sum}
AgSummary<-AgDat %>% 
  summarize(Total_F = sum(P_t),
            Alpha=min(t), Omega=max(t), Duration=max(t)/max(t),
            Mu=round(
              sum(t*P_t),
              1),
            Var=round(
              sum((t-Mu)^2*P_t),
              3),
            CS=round(
              sum((t-Mu)^3*P_t)/sqrt(Var)^3,
              1),
            CK=round(
              sum((t-Mu)^4*P_t)/sqrt(Var)^4-3,
              1)
            )
```

**Table 1**. Comparison of flowering schedule summary statistics

```{r ag_table}
kable(cbind(t(SumStats),c("Aggregate",t(AgSummary))), format = "markdown")
```

# Glossary

## Individual summary statistics

Characteristics of flowering schedules for the $i_{th}$ individual...

* Proportion ($P$) of flowers ($F$) produced at time $t$: $P_{i,t} = \frac{F_{i,t}}{\sum_t F_{i,t}}$
* Day of first flower: $\alpha_i = \min\{t \mid F_{i,t} > 0\}$ 
* Day of last flower: $\Omega_i = \max\{t \mid F_{i,t} > 0\}$ 
* Duration of flowering: $\Omega_i - \alpha_i$  
* Mean day of flower: $\mu_i = \sum_t t P_{i,t}$
* Flowering variance: $\sigma_i^2 = \sum_t (t-\mu)^2 P_t$
* Flowering coefficient of skewness: $CS_i = \frac{\sum_t (t-\mu_i)^3 P_{i,t}}{\sigma_i^3}$
* Flowering coefficient of kurtosis: $CK_i = \frac{\sum_t (t-\mu_i)^4 P_{i,t}}{\sigma_i^4}-3$

## Population *averaged* summary statistics

Averaging individual summary statistics for the $j_{th}$ population...

* Number of individuals: $N_j$ 
* Average day of first flower: $\bar \alpha_j = \frac{\sum_i \alpha_{i,j}}{N_j}$
* Average day of last flower: $\bar \Omega_j = \frac{\sum_i \Omega_{i,j}}{N_j}$
* Average duration of flowering: $\bar \Omega_j - \bar \alpha_j = \frac{\sum_i (\alpha_{i,j}-\Omega_{i,j})}{N_j} = \frac{\sum_i (\alpha_{i,j})}{N_j} - \frac{\sum_i \Omega_{i,j}}{N_j}$
* Average mean day of flower: $\bar \mu_j = \frac{\sum_i^N \mu_{i,j}}{N_j}$
* Average flowering variance: $\bar \sigma_j = \frac{\sum_i^N \sigma_{i,j}}{N_j}$
* Variance of mean day of flower: $\bar \sigma_j = \frac{\sum_i^N (\sigma_{i,j} - \bar \sigma_j)}{N_j}$
* Variance of flowering variance: $\bar \sigma_j = \frac{\sum_i^N (\sigma_{i,j} - \bar \sigma_j)}{N_j}$

## Population *aggregated* summary statistics

Aggregating observations across individuals for the $j_{th}$ population...

* Aggregated number of flowers at time $t$: $F_{t,j}^\sum = \sum_i F_{i,t,j}$
* Aggregated proportion of flowers at time $t$: $P_{t,j}^\sum = \frac{F_{t,j}^\sum}{\sum_t F_{t,j}^\sum} = \frac{\sum_i F_{i,t,j}}{\sum_t \sum_i F_{i,t,j}}$
* Aggregated day of first flower: $\alpha_j^\sum = \min \{t \mid \sum_i F_{i,t,j} > 0\}$
* Aggregated day of last flower: $\Omega_j^\sum = \max \{t \mid \sum_i F_{i,t,j} > 0\}$
* Aggregated duration of flower: $\Omega_j^\sum - \alpha_j^\sum$
* Aggregated mean day of flower: $\mu_j^\sum = \sum_t tP_{t,j}^\sum = \sum_t t \frac{\sum_i F_{i,t,j}}{\sum_t \sum_i F_{i,t,j}}$
* Aggregated flowering variance: $\sigma_j^{2\sum} = \sum_t (t-\mu_j^\sum)^2 P_{t,j}^\sum = \sum_t (t-\sum_t t \frac{\sum_i F_{i,t,j}}{\sum_t \sum_i F_{i,t,j}})^2 \frac{\sum_i F_{i,t,j}}{\sum_t \sum_i F_{i,t,j}}$
