==> AkbarEtAl_LythrumFloralDisplay/2_MomentsandEmergentProperties.Rmd <==
---
title: "Calculating the Central Moments & Emergent Properties"
author: "Mia Akbar"
date: "`r Sys.Date()`"
output: pdf_document
---

```{r setup}
library(tidyr)
library(dplyr)
library(ggplot2)
source("http://bit.ly/theme_pub")
theme_set(theme_pub()) 
setwd("./")
ifelse((!dir.exists(file.path("./", "Figures"))), (dir.create(file.path("./", "Figures"))), FALSE)
```

# Prepare data

Relative floral display data, seed family/population information and additional traits 
```{r raw_data}
Data <- read.csv("./Data/LythrumDisplay2008.csv")
names(Data)[grep("Num",names(Data))]<-"Ind"
str(Data)

# Veg: Final vegetative height
# PC1: Position on axis characterizing FTS trade-off
# HinfW: Inflorescence biomass at harvest
```

Remove problematic individuals 

```{r cancelled_guys}
Data <- Data %>% 
  filter(TotFlwr > 10) %>% 
  # Display too small to meaningfully characterize schedule shape
  filter(Ind != 187 & Ind != 440) 
# Start day not recorded 
```

Long format of floral display data

```{r display_long}
Disp <- Data %>% 
  pivot_longer(d0:d60, 
               names_to = "RelDay", 
               values_to = "OpFlwr") %>% 
  mutate(RelDay = as.numeric(gsub("d","",RelDay)),
         AbsDay = RelDay + Start) %>% 
  select(Ind, Fam, Pop, Lat, AbsDay, OpFlwr)
```

```{r}
write.csv(Disp, "./Data/Disp.csv", row.names = FALSE)
```


Plant family and population information

```{r pop_info}
Pops <- Data %>% 
  select(Ind, Fam, Pop, Lat)
```

```{r export_popinfo}
write.csv(Pops, "./Data/PopInfo.csv", row.names = FALSE)
```


# Key Functions

Basic statistics:

First day that at least one flower is observed

```{r start_function}
startFlwr<-function(Days=NA, NF=NA){return(
  min(Days[NF>0],na.rm=T)
)}
```

Last day that at least one flower is observed

```{r end_function}
endFlwr<-function(Days=NA, NF=NA){return(
  max(Days[NF>0],na.rm=T)
)}
```

Duration of flowering

```{r duration_function}
FlwrDuration<-function(Days=NA, NF=NA){return(
  endFlwr(Days,NF)-startFlwr(Days,NF)
)}
```


Given a vector NF of flower numbers by day:

1. Vector of proportion of flowers

```{r Flwr}
pFlwr<-function(NF=NA){return(
  NF/sum(NF, na.rm=T)
  )}
```

2. Vector of days weighted by the proportion of flowers on each day

```{r WeightDay}
WeightDay<-function(Days=NA, NF=NA){return(
  Days*pFlwr(NF)
  )}
```

3. Weighted mean flowering day

```{r MeanDay}
MeanDay<-function(Days=NA, NF=NA){return(
  sum(WeightDay(Days,NF),na.rm=T)
)}
```

4. Weighted variance

```{r Var}
VarDay<-function(Days=NA, NF=NA){return(
  sum((Days-MeanDay(Days,NF))^2*pFlwr(NF),na.rm=T)
)}
```

5. Coefficient of variance

```{r CV}
CoefVarDay<-function(Days=NA, NF=NA){return(
  VarDay(Days,NF)/MeanDay(Days,NF)
)}
```

6. Coefficient of skewness

```{r Cskew}
CoefSkewDay<-function(Days=NA, NF=NA){return(
  sum((Days-MeanDay(Days,NF))^3*pFlwr(NF),na.rm=T)/
    sqrt(VarDay(Days,NF))^3
)}
```

7. Coefficient of excess kurtosis

```{r Ckurt}
CoefKurtDay<-function(Days=NA, NF=NA){return(
  sum((Days-MeanDay(Days,NF))^4*pFlwr(NF),na.rm=T)/
    sqrt(VarDay(Days,NF))^4 - 3
)}
```


# Summarize data

## For Loop

Summary stats for each individual

```{r ind_calcs}
Moments<-data.frame(Ind=NA,Start=NA, Mean=NA,
                    Var=NA, CV=NA, 
                    Skew=NA, Kurtosis=NA,
                    End=NA, Duration=NA)
Ind<-unique(Disp$Ind)
for(i in 1:length(Ind)){
  Days<-Disp$AbsDay[Disp$Ind==Ind[i]] # Subset days for individual i
  NF<-Disp$OpFlwr[Disp$Ind==Ind[i]] # Subset flowers for individual i
  Moments[i,]<-c(Ind[i],startFlwr(Days,NF),MeanDay(Days,NF),
                 VarDay(Days,NF),CoefVarDay(Days,NF),
                 CoefSkewDay(Days,NF),CoefKurtDay(Days,NF),
                 endFlwr(Days,NF),FlwrDuration(Days,NF))
  # Clear objects from memory
  Days<-NA
  NF<-NA
}
#Moments$Ind<-as.factor(Moments$Ind)
```

Summary stats for population aggregates

```{r aggregate_calcs}
AgDisp<- Disp %>%
  select(!c("Ind","Fam")) %>%
  group_by(Pop,Lat,AbsDay) %>%
  summarize(OpFlwr=sum(OpFlwr,na.rm=T),
            Lat=mean(Lat,na.rm=T))

AgMoments<-data.frame(Lat=NA,Start=NA, Mean=NA,
                    Var=NA, CV=NA, 
                    Skew=NA, Kurtosis=NA,
                    End=NA, Duration=NA)
Lat<-unique(AgDisp$Lat)
for(i in 1:length(Lat)){
  Days<-AgDisp$AbsDay[AgDisp$Lat==Lat[i]] # Subset days for population i
  NF<-AgDisp$OpFlwr[AgDisp$Lat==Lat[i]] # Subset flowers for population i
  AgMoments[i,]<-c(Lat[i],startFlwr(Days,NF),MeanDay(Days,NF),
                 VarDay(Days,NF),CoefVarDay(Days,NF),
                 CoefSkewDay(Days,NF),CoefKurtDay(Days,NF),
                 endFlwr(Days,NF),FlwrDuration(Days,NF))
  # Clear objects from memory
  Days<-NA
  NF<-NA
}
```

```{r export_moments}
write.csv(Moments, "./Data/Moments.csv", row.names = FALSE)
```

# Individual  Correlations

```{r ind_corr}
round(cor(Moments[,-1]),2)
```

# Population Stats

## Population Average

```{r pop_mean_calcs}
PopStats<-left_join(Moments,Pops,by="Ind") %>%
  select(!c(Ind,Fam,Pop)) %>%
  pivot_longer(cols=Start:Duration,names_to="Stat",values_to="Value") %>%
  group_by(Lat,Stat) %>% 
  summarize(MeanValue=mean(Value))
  
```

```{r popmean_corr}
PopStatsWide<-pivot_wider(PopStats,names_from="Stat",values_from="MeanValue")
round(cor(PopStatsWide[,-1]),2)
```

```{r popmean_graphs, fig.height=8,fig.width=6}
PopStats$Stat<-factor(PopStats$Stat, 
                     levels=c("Start","End","Duration","Mean",
                              "Var","CV","Skew","Kurtosis"))
ggplot(aes(x=Lat,y=MeanValue),data=PopStats) + 
  geom_point() + facet_wrap(vars(Stat),scales="free") + 
  geom_smooth(method="lm",formula = y ~ x + I(x^2))
```

## Population Aggregate

```{r aggregate_corr}
round(cor(AgMoments[,-1]),2)
```

```{r aggregate_graphs, fig.height=8,fig.width=6}
AgStats<-pivot_longer(AgMoments,cols=Start:Duration,
                      names_to="Stat",values_to="AgValue")
AgStats$Stat<-factor(AgStats$Stat, 
                     levels=c("Start","End","Duration","Mean",
                              "Var","CV","Skew","Kurtosis"))
ggplot(aes(x=Lat,y=AgValue),data=AgStats) + 
  geom_point() + facet_wrap(vars(Stat),scales="free") + 
  geom_smooth(method="lm",formula = y ~ x + I(x^2))
```

```{r export_moments}
write.csv(AgMoments, "./Data/AgMoments.csv", row.names = FALSE)
```

## Emergent Properties

Note some values can be negative or positive (e.g., skew) so log-ratios don't work. Tried difference in z-score but then the problem is the average emergence is not informative (will be zero by definition). Decided to go with raw difference, which is easier to interpret biologically.

```{r emergent_calcs}
ESize<-function(x_pop=NA,y_ag=NA){return(
  (y_ag-x_pop)
  )}
Emergent<-data.frame(Lat=PopStatsWide$Lat,
                     Start=ESize(PopStatsWide$Start,AgMoments$Start),
                     Mean=ESize(PopStatsWide$Mean,AgMoments$Mean),
                     Var=ESize(PopStatsWide$Var,AgMoments$Var),
                     CV=ESize(PopStatsWide$CV,AgMoments$CV),
                     Skew=ESize(PopStatsWide$Skew,AgMoments$Skew),
                     Kurtosis=ESize(PopStatsWide$Kurtosis,AgMoments$Kurtosis),
                     End=ESize(PopStatsWide$End,AgMoments$End),
                     Duration=ESize(PopStatsWide$Duration,AgMoments$Duration))
```

```{r emergent_corr}
round(cor(Emergent[,-1]),2)
```

```{r emergent_graphs, fig.height=8,fig.width=6}
EmStats<-pivot_longer(Emergent,cols=Start:Duration,
                      names_to="Stat",values_to="EmScore")
EmStats$Stat<-factor(EmStats$Stat, 
                     levels=c("Start","End","Duration","Mean",
                              "Var","CV","Skew","Kurtosis"))
ggplot(aes(x=Lat,y=EmScore),data=EmStats) + 
  geom_point() + facet_wrap(vars(Stat),scales="free") + 
  geom_smooth(method="lm",formula = y ~ x + I(x^2))
```

```{r export_emergentproperties}
write.csv(EmStats, "./Data/Emergent.csv", row.names = FALSE)
```
