==> AkbarEtAl_LythrumFloralDisplay/7_TraitCorrelations.Rmd <==
---
title: "Visualizing Correlations of Flowering Schedule Characteristics"
author: "Mia Akbar"
date: "`r Sys.Date()`"
output: pdf_document
---

```{r setup}
library(tidyr)
library(dplyr)
library(ggplot2)
library(viridis)
library(corrplot)


theme_custom<-function(base_size = 24, base_family = ""){
  theme_bw(base_size = base_size, base_family = base_family) %+replace% 
    theme(axis.text = element_text(size = rel(0.8),colour="black"),
          axis.ticks = element_line(colour = "black"), 
          axis.title = element_text(size = rel(1.5), colour = "black"),
          legend.position = "none",
          panel.border = element_rect(fill = NA, colour = NA), 
          panel.grid.major = element_line(colour = NA,size = 0), 
          panel.grid.minor = element_line(colour = NA,size = 0), 
          axis.line = element_line(colour ="black")
    )
}

setwd("./")
ifelse((!dir.exists(file.path("./", "Figures"))), (dir.create(file.path("./", "Figures"))), FALSE)
```

# Prepare data

```{r bootgraph_data}
BootOut<-read.csv("./Data/BootOut.csv",header=T)

BootMeans<-dplyr::select(BootOut,Lat,Stat,Mean=mMean,Aggregate=agMean) %>%
  pivot_longer(cols=Mean:Aggregate,names_to="Calc_Type",values_to="BootMean")
BootLow<-dplyr::select(BootOut,Lat,Stat,Mean=mCIlow,Aggregate=agCIlow) %>%
  pivot_longer(cols=Mean:Aggregate,names_to="Calc_Type",values_to="CI_low")
BootUp<-dplyr::select(BootOut,Lat,Stat,Mean=mCIup,Aggregate=agCIup) %>%
  pivot_longer(cols=Mean:Aggregate,names_to="Calc_Type",values_to="CI_up")

GraphDat<-full_join(BootMeans,BootLow,by=c("Lat","Stat","Calc_Type")) %>%
  full_join(BootUp,by=c("Lat","Stat","Calc_Type"))
```

```{r bootavg_wide}
BootMeansWide <- BootMeans %>% 
  filter(Calc_Type == "Mean") %>% 
  pivot_wider(names_from = Stat, values_from = BootMean) %>% 
  dplyr::select(!Calc_Type)
```

```{r pcoa_axes}
PCoAAxes <- read.csv("./Data/PCoAAxes.csv",header=T) %>% rename(Ind = Num)
PCoAMeans <- PCoAAxes %>% 
  group_by(Pop,Lat) %>% 
  summarise(PCoA1 = mean(PCoA1),
            PCoA2 = mean(PCoA2),
            PCoA3 = mean(PCoA3),
            PCoA4 = mean(PCoA4))
BootMeansWide <- left_join(BootMeansWide,PCoAMeans) %>% 
  dplyr::select(!c(Pop, CV, PCoA3, PCoA4, End))
```

```{r pop_info}
PopInfo <- read.csv("./Data/PopInfo.csv") %>% rename(Ind=Num)
IndMoments <- read.csv("./Data/Moments.csv",header=T)
IndMoments <- left_join(IndMoments, PopInfo) %>% 
  dplyr::select(!c(Fam,Pop,CV)) %>% left_join(PCoAAxes) %>% dplyr::select(!c(Pop,PCoA3,PCoA4,End))
```

```{r ind_deviations}
Ind <- IndMoments %>% dplyr::select(Ind,Lat)
BootMeansInd <- left_join(Ind,BootMeansWide, by="Lat") ##for deviations
```

```{r agboot_wide}
AgMoments <- read.csv("./Data/AgMoments.csv")
```

# Visualize 

## Ind 

```{r popind_mat}
IndMomentsCorr <- IndMoments %>% left_join(PCoAAxes) %>% dplyr::select(Start,Duration,Mean,Var,Skew,Kurtosis, PCoA1, PCoA2)
IndMat <- round(cor(IndMomentsCorr),2)
```

```{r popind_plot}
# test for significant correlations
TestInd <- cor.mtest(IndMomentsCorr, conf.level = 0.95)

# plot 
pdf("./Figures/CorrelationPlots/Ind.pdf",width=6.5,height=6.5)
corrplot(IndMat, p.mat = TestInd$p, method = 'circle', type = 'lower', addCoef.col = "black", insig='blank', tl.col = 'black', tl.srt = 45, number.font=3, col=colorRampPalette(c("#fcfdbf","#fc8961","#b73779"))((200)), diag = FALSE)$corrPos -> p1
text(p1$x, p1$y, round(p1$corr, 2), font=3) 
dev.off()
p1 <- NULL # clean up 
```

## Pop Aggregate

```{r popag_mat}
AgMoments <- AgMoments %>%dplyr::select(!Lat)
AgMat <- round(cor(AgMoments),2)
```

```{r popag_plot}
# test for significant correlations
TestAg <- cor.mtest(AgMoments, conf.level = 0.95)

# plot 
pdf("./Figures/CorrelationPlots/Agg.pdf",width=6.5,height=6.5)
corrplot(AgMat, p.mat = TestAg$p, method = 'circle', type = 'lower', addCoef.col = "black", insig='blank', tl.col = 'black', tl.srt = 45, number.font=3, col=colorRampPalette(c("#fcfdbf","#fc8961","#b73779"))((200)), diag = FALSE)$corrPos -> p1
text(p1$x, p1$y, round(p1$corr, 2), font=3) 
dev.off()
p1 <- NULL # clean up 
```

## Pop Average 

```{r popavg_mat}
BootMeansWide <- BootMeansWide %>%dplyr::select(!Lat)
AvgMat <- round(cor(BootMeansWide),2)
```

```{r popavg_plot}
# test for significant correlations
TestAvg <- cor.mtest(BootMeansWide, conf.level = 0.95)

# plot 
pdf("./Figures/CorrelationPlots/Avg.pdf",width=6.5,height=6.5)
corrplot(AvgMat, p.mat = TestAvg$p, method = 'circle', type = 'lower', addCoef.col = "black", insig='blank', tl.col = 'black', tl.srt = 45, number.font=3, col=colorRampPalette(c("#fcfdbf","#fc8961","#b73779"))((200)), diag = FALSE)$corrPos -> p1
text(p1$x, p1$y, round(p1$corr, 2), font=3) 
dev.off()
p1 <- NULL # clean up 
```

## Ind after accounting for divergence among populations

```{r deviations}
Deviations <- IndMoments
for (trait in colnames(IndMoments)) {
  if (!(trait %in% c("Ind", "Lat"))) {
  Deviations[[trait]] <- IndMoments[[trait]] - BootMeansInd[[trait]]
  }
}

Deviations <- Deviations %>% dplyr::select(!c(Ind,Lat))
```

```{r deviation_mat}
DevMat <- round(cor(Deviations),2)
DevMat
```

```{r deviation_plot}
# test for significant correlations
TestDev <- cor.mtest(Deviations, conf.level = 0.95)

# plot 
pdf("./Figures/CorrelationPlots/IndDev.pdf",width=6.5,height=6.5)
corrplot(DevMat, p.mat = TestDev$p, method = 'circle', type = 'lower', addCoef.col = "black", insig='blank', tl.col = 'black', tl.srt = 45, number.font=3, col=colorRampPalette(c("#fcfdbf","#fc8961","#b73779"))((200)), diag = FALSE)$corrPos -> p1
text(p1$x, p1$y, round(p1$corr, 2), font=3) 
dev.off()
p1 <- NULL # clean up 
```
